Change geometry to use high performance static octree
